pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "venustracker"
version = "0.1.0"
description = "Venus-focused astronomical tracker powered by JPL Horizons, VizieR, COBS, and DSS"
requires-python = ">=3.10"
dependencies = [
    "astropy>=6.0",
    "astroquery>=0.4.7",
    "httpx>=0.27",
    "pandas>=2.0",
    "numpy>=1.26",
    "typer[all]>=0.12",
    "rich>=13.0",
    "diskcache>=5.6",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
vt = "venustracker.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/venustracker"]

[tool.ruff]
line-length = 100
target-version = "py310"
